[fp-rs] Converted bluetooth module into standalone crate, with fastpair UI provided as an example.
