+++ fastpair/rust/bluetooth/Cargo.toml
@@ -13,18 +13,21 @@
 # limitations under the License.
 
 [package]
-name = "fastpair"
+name = "bluetooth"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-futures = { version = "0.3", features = ["executor"] }
+futures = { version = "0.3" }
 tracing = "0.1.37"
 cfg-if = "1.0.0"
 async-trait = "0.1"
 thiserror = "1.0.43"
+
+[dev-dependencies]
+futures = { version = "0.3", features = ["executor"] }
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.48", features = [
