+++ fastpair/rust/bluetooth/src/common/advertisement.rs
