+++ fastpair/rust/bluetooth/src/common/error.rs
