+++ fastpair/rust/bluetooth/src/common/mod.rs
