+++ fastpair/rust/bluetooth/src/unsupported/mod.rs
