+++ fastpair/rust/bluetooth/tests/integration_test.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use fastpair::*;
+use bluetooth::*;
 
 mod tests {
     use super::*;
